apps/analytics-worker/requirements.txt
celery[redis]==5.3.6
redis==5.0.1
psycopg2-binary==2.9.9
pymysql==1.1.0
pymongo==4.6.1
structlog==24.1.0
